+++ DataCollector/src/main/java/com/braintreepayments/api/datacollector/DataCollectorResult.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.datacollector
 
 /**
  * Result of collecting device data for fraud detection