+++ LocalPayment/src/main/java/com/braintreepayments/api/LocalPaymentClient.java
@@ -7,6 +7,7 @@ import android.net.Uri;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 
+import com.braintreepayments.api.datacollector.DataCollector;
 import com.braintreepayments.api.sharedutils.Json;
 
 import org.json.JSONException;